rename camp file name
