+++ backend/typescript/models/campSession.model.ts
